requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.7
django-allauth==0.52.0
django-cacheops==6.2
django-cors-headers==3.13.0
django-countries==7.5.1
django-extensions==3.2.1
django-guardian==2.4.0
django-phonenumber-field==7.0.2
django-session-timeout==0.1.0
django-simple-history==3.2.0
django-sms==0.6.0
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-guardian==0.3.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-standardized-errors==0.12.5
drf-tracking==1.5.0
drf-writable-nested==0.7.0
executing==1.2.0
funcy==1.18
guardian==0.2.2
idna==3.4
ipython==8.10.0
isort==5.12.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
json-log-decorator==2.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.6
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.37
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
python-dotenv==0.21.1
python-json-logger==2.0.1
python-logging-loki==0.3.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339==6.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
